[U] Chunk loaders have a common base class now
