--- VoxeUnity/Assets/Voxelmetric/Code/Utilities/ChunkLoaders/IChunkLoader.cs
@@ -1,13 +1,0 @@
-﻿using Voxelmetric.Code.Core;
-
-namespace Voxelmetric.Code.Utilities.ChunkLoaders
-{
-    public interface IChunkLoader
-    {
-        void PreProcessChunks();
-        void ProcessChunks();
-        void PostProcessChunks();
-
-        void ProcessChunk(Chunk chunk);
-    }
-}
