--- VoxeUnity/Assets/Voxelmetric/Code/Utilities/ChunkLoaders/IChunkLoader.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b0eb5eaea2a6a944093be4731eb51a8b
-timeCreated: 1466242757
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
